# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  internal_handlers.CreateRequirementAPIRequest:
    properties:
      description:
        type: string
      due_date:
        type: string
      max_report_age_days:
        type: integer
      minimum_grade:
        type: string
      passing_score:
        type: integer
      priority:
        type: string
      questionnaire_id:
        type: string
      relationship_id:
        type: string
      title:
        type: string
      type:
        type: string
    required:
    - relationship_id
    - title
    - type
    type: object
...
  internal_handlers.ErrorResponse:
    properties:
      error:
        type: string
      message:
        type: string
    type: object
...
  internal_handlers.PaginatedRequirementsResponse:
    properties:
      items:
        items:
          $ref: '#/definitions/internal_handlers.RequirementResponse'
        type: array
      limit:
        type: integer
      page:
        type: integer
      total_count:
        type: integer
      total_pages:
        type: integer
    type: object
...
  internal_handlers.RequirementResponse:
    properties:
      assigned_at:
        type: string
      company_id:
        type: string
      created_at:
        type: string
      days_until_due:
        type: integer
      description:
        type: string
      due_date:
        type: string
      id:
        type: string
      is_overdue:
        type: boolean
      max_report_age_days:
        type: integer
      minimum_grade:
        type: string
      passing_score:
        type: integer
      priority:
        type: string
      questionnaire_id:
        type: string
      relationship_id:
        type: string
      status:
        type: string
      status_history:
        items:
          $ref: '#/definitions/internal_handlers.RequirementStatusChangeResp'
        type: array
      supplier_id:
        type: string
      title:
        type: string
      type:
        type: string
      updated_at:
        type: string
    type: object
...
  internal_handlers.RequirementStatusChangeResp:
    properties:
      changed_at:
        type: string
      from_status:
        type: string
      reason:
        type: string
      to_status:
        type: string
    type: object
  internal_handlers.ReviewActionRequest:
    properties:
      notes:
        type: string
      reason:
        type: string
    type: object
  internal_handlers.ReviewResponseDetails:
    properties:
      grade:
        type: string
      id:
        type: string
      is_reviewed:
        type: boolean
      is_submitted:
        type: boolean
      max_score:
        type: integer
      passed:
        type: boolean
      review_notes:
        type: string
      reviewed_at:
        type: string
      score:
        type: integer
      started_at:
        type: string
      submitted_at:
        type: string
    type: object
  internal_handlers.ReviewSubmissionDetails:
    properties:
      answers:
        items:
          $ref: '#/definitions/internal_handlers.SubmissionAnswerResponse'
        type: array
      completion_time_minutes:
        type: integer
      id:
        type: string
      max_possible_score:
        type: integer
      must_pass_failed:
        type: boolean
      passed:
        type: boolean
      percentage_score:
        type: number
      topic_scores:
        items:
          $ref: '#/definitions/internal_handlers.TopicScoreResponse'
        type: array
      total_score:
        type: integer
    type: object
  internal_handlers.ReviewSubmissionResponse:
    properties:
      requirement:
        $ref: '#/definitions/internal_handlers.RequirementResponse'
      response:
        $ref: '#/definitions/internal_handlers.ReviewResponseDetails'
      submission:
        $ref: '#/definitions/internal_handlers.ReviewSubmissionDetails'
    type: object
...
  internal_handlers.SubmissionAnswerResponse:
    properties:
      is_must_pass_met:
        type: boolean
      max_points:
        type: integer
      points_earned:
        type: integer
      question_id:
        type: string
      selected_options:
        items:
          type: string
        type: array
      text_answer:
        type: string
    type: object
...
  internal_handlers.TopicScoreResponse:
    properties:
      max_score:
        type: integer
      percentage_score:
        type: number
      score:
        type: integer
      topic_id:
        type: string
      topic_name:
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@nisfix.io
    name: NisFix Support
  description: B2B Supplier Security Portal API - Companies manage supplier security
    requirements through questionnaires and CheckFix reports
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: NisFix Backend API
  version: "1.0"
paths:
...
  /requirements:
    get:
      consumes:
      - application/json
      description: Lists all requirements for the company
      parameters:
      - description: Filter by status
        in: query
        name: status
        type: string
      - description: Filter by relationship
        in: query
        name: relationship_id
        type: string
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 20
        description: Items per page
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_handlers.PaginatedRequirementsResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_handlers.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/internal_handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: List requirements
      tags:
      - Requirements
    post:
      consumes:
      - application/json
      description: Creates a new requirement for a supplier
      parameters:
      - description: Create request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_handlers.CreateRequirementAPIRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/internal_handlers.RequirementResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_handlers.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_handlers.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/internal_handlers.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/internal_handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Create a requirement
      tags:
      - Requirements
...
  /requirements/{id}/approve:
    post:
      consumes:
      - application/json
      description: Approves a submitted requirement
      parameters:
      - description: Requirement ID
        in: path
        name: id
        required: true
        type: string
      - description: Approval notes
        in: body
        name: request
        schema:
          $ref: '#/definitions/internal_handlers.ReviewActionRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_handlers.RequirementResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_handlers.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_handlers.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/internal_handlers.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/internal_handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Approve requirement
      tags:
      - Review
...
  /requirements/{id}/reject:
    post:
      consumes:
      - application/json
      description: Rejects a submitted requirement
      parameters:
      - description: Requirement ID
        in: path
        name: id
        required: true
        type: string
      - description: Rejection reason
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_handlers.ReviewActionRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_handlers.RequirementResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_handlers.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_handlers.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/internal_handlers.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/internal_handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Reject requirement
      tags:
      - Review
  /requirements/{id}/request-revision:
    post:
      consumes:
      - application/json
      description: Requests revision for a submitted requirement
      parameters:
      - description: Requirement ID
        in: path
        name: id
        required: true
        type: string
      - description: Revision reason
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_handlers.ReviewActionRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_handlers.RequirementResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_handlers.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_handlers.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/internal_handlers.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/internal_handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Request revision
      tags:
      - Review
  /requirements/{id}/review:
    get:
      consumes:
      - application/json
      description: Gets the submission details for reviewing
      parameters:
      - description: Requirement ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_handlers.ReviewSubmissionResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_handlers.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/internal_handlers.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/internal_handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get submission for review
      tags:
      - Review
...
securityDefinitions:
  BearerAuth:
    description: 'Enter your bearer token in the format: Bearer {token}'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"